renames + mongo uri config
