+++ packages/catalog-process/.env
@@ -3,3 +3,4 @@ PORT=3000
 POSTGRESQL_URI="postgresql://root:root@localhost:6001/root?schema=root"
 LOG_LEVEL=info
 MOCK_FILE_MANAGER=true
+MONGO_URI="mongodb://root:example@localhost:27017"
